Finished prep
Identified main issue, which was a cause of issues at first, namely, the interpolation of the waveform parameters originally in t_align. This is where most of the difference is coming from. Will investigate in the analysis file if doing this when necessary makes a large difference. (and whether it is in fact the issue since this is an expectation.)