Savefile does not contain 0 ecc
Ading the zero eccentricity simulation fits into the file made no sense and ruined the interpolation, still artifacts going on and that process should be coded better (one loop instead of multiple). Priority now however is identifying the misbehaving points and seeing why they misbehave. It looks like the amplitude fits are decently well behaved but the omega fits are not as expected.